in_sdl.c: use a different approach to ignore mouse events.
We now use SDL_SetEventFilter() instead of SDL_PumpEvents().
This seems to fix a freeze reported by "galtgendo". Since
none of the uhexen2 developers have been able to reproduce
this freeze it's hard to say exactly what was wrong, but so
far this seems to fix it (code adapted from quakespasm).


git-svn-id: https://svn.code.sf.net/p/uhexen2/code/trunk@5197 eaf649e6-6c13-0410-b194-84221022858e